fix after merge

+++ src/shared/adaptations/adaptation.h
@@ -109,7 +109,7 @@ class ParticleWithLocalRefinement : public SPHAdaptation
 {
   public:
     Real *h_ratio_; /**< the ratio between reference smoothing length to variable smoothing length */
-    size_t *level_; /**< the mesh level of the particle */
+    int *level_;    /**< the mesh level of the particle */
 
     ParticleWithLocalRefinement(Real resolution_ref, Real h_spacing_ratio_, Real system_refinement_ratio, int local_refinement_level);
     virtual ~ParticleWithLocalRefinement(){};
